refactor(image-editor): encapsulate isImageFile in image-editor lib
Move the image-detection predicate out of libs/utils and into the image-editor
lib (alongside enrichEditedImage), exported from its public API. Groups the
image-ness logic with its owner and leaves libs/utils untouched.

Drop the file-extension fallback: rely on the authoritative isImage flag and
the image/* content type, both reliably present on the metadata the gate runs
against (dotAsset assetMetaData and enriched Binary temp files). File names are
not a trusted signal.

Refs #36363
