+++ core-web/libs/image-editor/src/index.ts
@@ -2,6 +2,7 @@ export { DotImageEditorComponent } from './lib/components/dot-image-editor/dot-i
 export * from './lib/models/image-editor.models';
 export { RANGES } from './lib/image-editor.constants';
 export { buildFilterChain, buildPreviewUrl, cleanUrl } from './lib/utils/image-filter-url.builder';
+export { isImageFile } from './lib/utils/is-image-file.util';
 export * from './lib/store/image-editor.events';
 export { ImageEditorStore } from './lib/store/image-editor.store';
 export { initialImageEditorState } from './lib/store/image-editor.state';
