+++ core-web/libs/utils/src/lib/shared/contentlet.utils.ts
@@ -72,67 +72,3 @@ export function checkMimeType(
 
     return false;
 }
-
-/**
- * Common image file extensions, used as a last-resort fallback when metadata
- * carries neither an explicit `isImage` flag nor a recognizable `image/*`
- * content type.
- */
-const IMAGE_FILE_EXTENSIONS = new Set([
-    'jpg',
-    'jpeg',
-    'png',
-    'gif',
-    'webp',
-    'bmp',
-    'svg',
-    'ico',
-    'tif',
-    'tiff',
-    'avif',
-    'heic',
-    'heif'
-]);
-
-/**
- * Whether a file name ends with a known image extension.
- *
- * @param name The file name to inspect.
- * @returns true when the name has a recognized image extension.
- */
-function hasImageExtension(name: string | undefined): boolean {
-    const extension = name?.split('.').pop()?.toLowerCase();
-
-    return !!extension && IMAGE_FILE_EXTENSIONS.has(extension);
-}
-
-/**
- * Determines whether a file's metadata describes an image, using layered signals
- * so detection stays correct even when the metadata is partial:
- *
- * 1. the authoritative server-computed `isImage` flag;
- * 2. an `image/*` content type;
- * 3. a known image file extension on the file name (last-resort fallback).
- *
- * Use this to gate image-only affordances (e.g. the image editor) uniformly
- * across Binary fields and the `dotAsset` referenced by Image/File fields. Pair
- * it with {@link getFileMetadata} to resolve the metadata first.
- *
- * @param metadata The resolved file metadata (may be partial, null or undefined).
- * @returns true when the metadata describes an image.
- */
-export function isImageFile(metadata: Partial<DotFileMetadata> | null | undefined): boolean {
-    if (!metadata) {
-        return false;
-    }
-
-    if (metadata.isImage) {
-        return true;
-    }
-
-    if (metadata.contentType?.toLowerCase().startsWith('image/')) {
-        return true;
-    }
-
-    return hasImageExtension(metadata.name);
-}
